rename TypedPropertiesKey and TypedPropertiesValue
