+++ sharding-orchestration/sharding-orchestration-center/sharding-orchestration-center-nacos/src/main/java/org/apache/shardingsphere/orchestration/center/instance/NacosPropertiesEnum.java
@@ -19,14 +19,14 @@ package org.apache.shardingsphere.orchestration.center.instance;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.underlying.common.properties.common.TypedPropertiesKey;
+import org.apache.shardingsphere.underlying.common.properties.common.TypedPropertyKey;
 
 /**
  * Nacos properties enum.
  */
 @RequiredArgsConstructor
 @Getter
-public enum NacosPropertiesEnum implements TypedPropertiesKey {
+public enum NacosPropertiesEnum implements TypedPropertyKey {
     
     /**
      * Nacos config service group name.
